RMPM_SC: Add typecast for CSL read and write function calls
The argument to CSL read and write is typecasted to uint32_t/ uintptr_t
The return value of CSL read is typecasted to read store type
This is in conformance with MISRA C 2012 Rule 11.4

+++ src/safety_checkers_rm.c
@@ -94,7 +94,7 @@ int32_t SafetyCheckers_rmGetRegCfg(uintptr_t *rmRegCfg, uint32_t size)
 	{
 		for (totalModule = 0U; totalModule < (sizeof(gSafetyCheckers_RmRegData) / sizeof(SafetyCheckers_RmRegData)); totalModule++)
 		{
-			totalLength += (gSafetyCheckers_RmRegData[totalModule].regNum) * (gSafetyCheckers_RmRegData[totalModule].regArrayLen);
+			totalLength += (uint32_t)((gSafetyCheckers_RmRegData[totalModule].regNum) * (gSafetyCheckers_RmRegData[totalModule].regArrayLen));
 		}
 		
 		if (totalLength > size)
@@ -111,7 +111,7 @@ int32_t SafetyCheckers_rmGetRegCfg(uintptr_t *rmRegCfg, uint32_t size)
 			{
 				for (regArrayLen = 0U; regArrayLen < gSafetyCheckers_RmRegData[totalModule].regArrayLen; regArrayLen++)
 				{
-					rmRegCfg[offset] = CSL_REG32_RD(gSafetyCheckers_RmRegData[totalModule].baseAddr +\
+					rmRegCfg[offset] = (uintptr_t)CSL_REG32_RD(gSafetyCheckers_RmRegData[totalModule].baseAddr +\
 					((gSafetyCheckers_RmRegData[totalModule].regStartOffset * regArrayLen) + gSafetyCheckers_RmRegData[totalModule].regOffsetArr[regNum]));
 
 					offset++;
@@ -142,7 +142,7 @@ int32_t SafetyCheckers_rmVerifyRegCfg (const uintptr_t *rmRegCfg, uint32_t size)
 	{
 		for (totalModule = 0U; totalModule < (sizeof(gSafetyCheckers_RmRegData) / sizeof(SafetyCheckers_RmRegData)); totalModule++)
 		{
-			totalLength += (gSafetyCheckers_RmRegData[totalModule].regNum) * (gSafetyCheckers_RmRegData[totalModule].regArrayLen);
+			totalLength += (uint32_t)((gSafetyCheckers_RmRegData[totalModule].regNum) * (gSafetyCheckers_RmRegData[totalModule].regArrayLen));
 		}
 
 		if (totalLength > size)
@@ -159,9 +159,9 @@ int32_t SafetyCheckers_rmVerifyRegCfg (const uintptr_t *rmRegCfg, uint32_t size)
 			{
 				for (regArrayLen = 0U; regArrayLen < gSafetyCheckers_RmRegData[totalModule].regArrayLen; regArrayLen++)
 				{
-					uint32_t readData = CSL_REG32_RD(gSafetyCheckers_RmRegData[totalModule].baseAddr +\
+					uint32_t readData = (uint32_t)CSL_REG32_RD(gSafetyCheckers_RmRegData[totalModule].baseAddr +\
 					((gSafetyCheckers_RmRegData[totalModule].regStartOffset * regArrayLen) + gSafetyCheckers_RmRegData[totalModule].regOffsetArr[regNum]));			
-					mismatchCnt |= rmRegCfg[offset] ^ readData;
+					mismatchCnt |= (uint32_t)(rmRegCfg[offset] ^ readData);
 					offset++;
 				}
 			}
